server/requirements.txt
langchain
openai
tiktoken
python-dotenv